+++ src/Vortice.XAudio2/IXAudio2Voice.cs
@@ -1,0 +1,14 @@
+// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using System;
+using SharpGen.Runtime;
+using Vortice.Multimedia;
+
+namespace Vortice.XAudio2
+{
+    public partial class IXAudio2Voice
+    {
+        public static XAudio2Version Version { get; internal set; }
+    }
+}